Fix race-condition issues when deploying new cards data
If there is a delay between the moment the new `version.json` is
served, and the moment the new `cards.json` is served, this will
confuse the extension which will save invalid data locally.

Instead, we now have cards.json and dictionary.json merged together
into `data.json`, along with a version field, to make sure the
received data is of the proper version.